feat(hw4): dockerize and document WIP
