+++ docs/homework4/README.md
@@ -1,0 +1,23 @@
+> Закончить систему
+
+Поправил `inventory` на `taskos` (исп.)
+
+> Если вы успели сделать всё в срок
+
+> Достать спроектированную систему из нулевой недели и сравнить её с текущим результатом. Найти места, которые вы сделали иначе, чем изначально планировали, что в текущем решении отличается от изначального, что лично вам понравилось из решений.
+
+В целом так же получилось, единственное я просто забыл о сервисе репортинга. Или впилил его прямо в биллинг, но там задача у него другая, так что выносить ок.
+
+> выписать идеи и подходы, которые вы хотели бы внедрить в ваших рабочих или других проектах, что бы мы могли их обсудить и подумать как лучше всего внедрить каждую из идей или подходов.
+
+У нас как-то витала в воздухе идея "а давайте всё-всё-всё сделаем ивент сурсингом - причём по причине "и тогда нам не нужны будут бекапы!". 
+
+Ну в общем понятно что её буду блочить. Там только в одном/двух специфичных местах ивент сурсинг сейчас нужен, и он там собственно есть в разных формах. А бекапы всё равно нужны, ёмаё!
+
+> добавить документацию, в которой расписать как ставить систему и как с ней работать
+
+yes
+
+> Причесать код и добавить к нему документацию там, где это следует сделать и записать небольшое видео (или сделать скриншоты) с тем, как система работает целиком. Это позволит вам показать репозиторий как пример кода для собеседований или поделиться им со знакомыми.
+
+not yet...
