+++ packages/accounting/src/main.ts
@@ -4,13 +4,13 @@ import { run as runReactions } from './reactions';
 import { getTotalStonksForDate } from './db';
 import { consumer, producer } from './kafka';
 import { makeAuthMiddleware, useCanRole } from '../../utils/src/lib/auth';
-import { ROLE_ACCOUNTANT, ROLE_ADMIN } from '@monorepo/kafka-users-common';
+import { ROLE_ACCOUNTANT, ROLE_ADMIN } from '@monorepo/utils';
 
 const app = express();
 
 app.use(bodyParser.json());
 
-const port = parseInt(process.env.PORT || '3333', 10);
+const port = parseInt(process.env.PORT || '3002', 10);
 const server = app.listen(port, async () => {
   await runReactions();
   await consumer.connect();
