+++ packages/accounting/src/reactions.ts
@@ -6,7 +6,7 @@ import {
   TASK_EVENT_ASSIGN,
   TASK_EVENT_COMPLETE,
   TaskEvent,
-} from '@monorepo/inventory-common/schema';
+} from '@monorepo/taskos-common/schema';
 import { match } from 'ts-pattern';
 import { penalty, reward } from './db';
 import { consumer, reportAccountsAggregate } from './kafka';
