--- packages/inventory-common/README.md
@@ -1,11 +1,0 @@
-# inventory-common
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build inventory-common` to build the library.
-
-## Running unit tests
-
-Run `nx test inventory-common` to execute the unit tests via [Jest](https://jestjs.io).
