+++ packages/taskos-common/README.md
@@ -1,0 +1,11 @@
+# taskos-common
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build taskos-common` to build the library.
+
+## Running unit tests
+
+Run `nx test taskos-common` to execute the unit tests via [Jest](https://jestjs.io).
