+++ packages/taskos-common/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@monorepo/inventory-common",
+  "name": "@monorepo/taskos-common",
   "version": "0.0.1",
   "type": "commonjs"
 }
