+++ packages/taskos-common/project.json
@@ -1,31 +1,31 @@
 {
-  "name": "inventory-common",
+  "name": "taskos-common",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/inventory-common/src",
+  "sourceRoot": "packages/taskos-common/src",
   "projectType": "library",
   "targets": {
     "build": {
       "executor": "@nx/js:tsc",
       "outputs": ["{options.outputPath}"],
       "options": {
-        "outputPath": "dist/packages/inventory-common",
-        "main": "packages/inventory-common/src/index.ts",
-        "tsConfig": "packages/inventory-common/tsconfig.lib.json",
-        "assets": ["packages/inventory-common/*.md"]
+        "outputPath": "dist/packages/taskos-common",
+        "main": "packages/taskos-common/src/index.ts",
+        "tsConfig": "packages/taskos-common/tsconfig.lib.json",
+        "assets": ["packages/taskos-common/*.md"]
       }
     },
     "lint": {
       "executor": "@nx/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["packages/inventory-common/**/*.ts"]
+        "lintFilePatterns": ["packages/taskos-common/**/*.ts"]
       }
     },
     "test": {
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "packages/inventory-common/jest.config.ts",
+        "jestConfig": "packages/taskos-common/jest.config.ts",
         "passWithNoTests": true
       },
       "configurations": {
