+++ packages/taskos/src/user/db.ts
@@ -1,4 +1,4 @@
 // TODO persist
-import { User, UserId } from '@monorepo/kafka-users-common';
+import { User, UserId } from '@monorepo/utils';
 
 export const users: Map<UserId, User> = new Map();
