+++ tsconfig.base.json
@@ -17,7 +17,7 @@
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "@monorepo/inventory-common/*": ["packages/inventory-common/src/lib/*"],
+      "@monorepo/taskos-common/*": ["packages/taskos-common/src/lib/*"],
       "@monorepo/kafka-users-common": [
         "packages/kafka-users-common/src/index.ts"
       ],
